Revise to accommodate running workflows with cwl-runner, and update README
